# PlantHereNodeJs/Presentation/PlantHere.WebAPI/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlantHere Nodejs",
    "version": "1.0"
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get Order By User Id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByUserIdQueryResultICollectionCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByUserIdQueryResultICollectionCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByUserIdQueryResultICollectionCustomResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "Update Order",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByIdQueryResultCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByIdQueryResultCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderByIdQueryResultCustomResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Create Order",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderCommandResultCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderCommandResultCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderCommandResultCustomResult"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Delete Order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              }
            }
          }
        }
      }
    },
    "/products/{uniqueId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get Product By Unique Id",
        "parameters": [
          {
            "name": "uniqueId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              }
            }
          }
        }
      }
    },
    "/products/{page}/{pageSize}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get Products By Page",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsByPageQueryResultIEnumerableCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsByPageQueryResultIEnumerableCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsByPageQueryResultIEnumerableCustomResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/products": {
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "Update Product",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByUniqueIdQueryResultCustomResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Create Product",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProductCommandResultCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProductCommandResultCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProductCommandResultCustomResult"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Delete Product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductCommandCustomResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddressQueryResult": {
        "type": "object",
        "properties": {
          "province": {
            "type": "string",
            "nullable": true
          },
          "district": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "line": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateOrderCommand": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressQueryResult"
          },
          "buyerId": {
            "type": "string",
            "nullable": true
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetOrderItemsQueryResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateOrderCommandResult": {
        "type": "object",
        "additionalProperties": false
      },
      "CreateOrderCommandResultCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateOrderCommandResult"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CreateProductCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "sellerId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateProductCommandResult": {
        "type": "object",
        "additionalProperties": false
      },
      "CreateProductCommandResultCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateProductCommandResult"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "DeleteProductCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DeleteProductCommandCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteProductCommand"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetOrderByIdQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressQueryResult"
          },
          "buyerId": {
            "type": "string",
            "nullable": true
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetOrderItemsQueryResult"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "GetOrderByIdQueryResultCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetOrderByIdQueryResult"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "GetOrderByUserIdQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressQueryResult"
          },
          "buyerId": {
            "type": "string",
            "nullable": true
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetOrderItemsQueryResult"
            },
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "discountedTotalPrice": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "GetOrderByUserIdQueryResultICollectionCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetOrderByUserIdQueryResult"
            },
            "nullable": true,
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "GetOrderItemsQueryResult": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discountedPrice": {
            "type": "number",
            "format": "double"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetProductByUniqueIdQueryResult": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discountedPrice": {
            "type": "number",
            "format": "double"
          },
          "uniqueId": {
            "type": "string",
            "nullable": true
          },
          "care": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetProductByUniqueIdQueryResultCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetProductByUniqueIdQueryResult"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "GetProductsByPageQueryResult": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discountedPrice": {
            "type": "number",
            "format": "double"
          },
          "uniqueId": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetProductsByPageQueryResultIEnumerableCustomResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetProductsByPageQueryResult"
            },
            "nullable": true,
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateOrderCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressQueryResult"
          },
          "buyerId": {
            "type": "string",
            "nullable": true
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetOrderItemsQueryResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateProductCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}